Add 'test' option for conditional execution
This option gives us the possibility to check the return code of a
command for a conditional execution of the given task. This way, we
could run grep, or anything else, to test if we need the task to be run.

When the test command returns true, ie 0, we avoid task execution,
otherwise, we run the task.